Bug fixes
Fixes a number of minor bugs and issues. Corrected the syntax for getting the size of a numpy array (i.e. when reporting out on the number of points in an elevation DEM); when comparing two numpy arrays, set the error threshold to a tolerance value greater than zero (DEM_DIFF_TOL) to account for numerical noise between different DEMs; when initially digitizing the elevation bands and HRUs added additional logic to account for and add HRUs when a new elevation band 'appears' at band index with an initial area of 0.